+++ helm/tests/ensure.py
@@ -14,6 +14,7 @@ LOGGER = logging.getLogger(__name__)
 service_monitor_name = "test-service-monitor"
 silence_name = "test-silence"
 cluster_name = "test-cluster"
+machinepool_name = "mp0"
 release_version = "20.0.0"
 cluster_apps_operator_version = "2.0.0"
 watch_label = "capi"
@@ -402,6 +403,8 @@ def azurecluster(kubernetes_cluster):
           labels:
             giantswarm.io/cluster: {cluster_name}
             cluster.x-k8s.io/cluster-name: {cluster_name}
+        spec:
+          location: ""
     """)
 
     kubernetes_cluster.kubectl("apply", input=c, output=None)
@@ -416,6 +419,49 @@ def azurecluster(kubernetes_cluster):
 
     kubernetes_cluster.kubectl(f"delete azurecluster {cluster_name}", output=None)
     LOGGER.info(f"AzureCluster {cluster_name} deleted")
+
+@pytest.fixture
+def azuremachinepool(kubernetes_cluster):
+    c = dedent(f"""
+        apiVersion: infrastructure.cluster.x-k8s.io/v1alpha4
+        kind: AzureMachinePool
+        metadata:
+          name: {machinepool_name}
+          namespace: default
+          labels:
+            "cluster.x-k8s.io/cluster-name": {cluster_name}
+            "giantswarm.io/cluster": {cluster_name}
+        spec:
+          identity: SystemAssigned
+          location: ""
+          strategy:
+            rollingUpdate:
+              deletePolicy: Oldest
+              maxSurge: 25%
+              maxUnavailable: 1
+            type: RollingUpdate
+          template:
+            osDisk:
+              diskSizeGB: 30
+              managedDisk:
+                storageAccountType: Premium_LRS
+              osType: Linux
+            sshPublicKey: ""
+            vmSize: Standard_D4s_v3
+    """)
+
+    kubernetes_cluster.kubectl("apply", input=c, output=None)
+    LOGGER.info(f"AzureMachinePool {machinepool_name} applied")
+
+    raw = kubernetes_cluster.kubectl(
+        f"get AzureMachinePool {machinepool_name}", output="yaml")
+
+    azuremachinepool = yaml.safe_load(raw)
+
+    yield azuremachinepool
+
+    kubernetes_cluster.kubectl(f"delete azuremachinepool {machinepool_name}", output=None)
+    LOGGER.info(f"AzureMachinePool {machinepool_name} deleted")
 
 
 # Silence fixtures
@@ -475,3 +521,37 @@ def silence_with_matchers(kubernetes_cluster):
 
     kubernetes_cluster.kubectl(f"delete silence {silence_name}", output=None)
     LOGGER.info(f"Silence {silence_name} deleted")
+
+@pytest.fixture
+def kubeadm_control_plane(kubernetes_cluster):
+    c = dedent(f"""
+        apiVersion: controlplane.cluster.x-k8s.io/v1alpha3
+        kind: KubeadmControlPlane
+        metadata:
+          labels:
+            cluster.x-k8s.io/cluster-name: {cluster_name}
+            cluster.x-k8s.io/watch-filter: capi
+          name: {cluster_name}
+          namespace: default
+        spec:
+          kubeadmConfigSpec:
+            clusterConfiguration:
+              controllerManager:
+                extraArgs:
+                  allocate-node-cidrs: "false"
+          infrastructureTemplate: {{}}
+          version: 1.22.0
+    """)
+
+    kubernetes_cluster.kubectl("apply", input=c, output=None)
+    LOGGER.info(f"KubeadmControlPlane {cluster_name} applied")
+
+    raw = kubernetes_cluster.kubectl(
+        f"get kubeadmcontrolplane {cluster_name}", output="yaml")
+
+    kcp = yaml.safe_load(raw)
+
+    yield kcp
+
+    kubernetes_cluster.kubectl(f"delete kubeadmcontrolplane {cluster_name}", output=None)
+    LOGGER.info(f"kubeadmcontrolplane {cluster_name} deleted")